move some functions from main.go to browse.go
